+++ www/src/redux/substores/student/studentStore.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 
 import resumeReviewReducer from './slices/resumeReviewSlice';
-import resumeReviewViewerReducer from './slices/resumeReviewViewer';
+import resumeReviewViewerReducer from './slices/resumeReviewViewerSlice';
 import uploadResumeReducer from './slices/uploadResumeSlice';
 
 const store = configureStore({
